Switch to reflection (now it's much cleaner)
