+++ build.gradle
@@ -1,54 +1,53 @@
 import org.apache.tools.ant.filters.ReplaceTokens
 
 plugins {
+    id 'java'
     id 'com.github.johnrengelman.shadow' version '5.1.0'
 }
 
-allprojects {
-    apply plugin: 'java'
-    apply plugin: 'com.github.johnrengelman.shadow'
-    group = 'net.leomixer17'
-    version = '1.3.9'
-    description = 'Create cool interactive books!'
-    sourceCompatibility = 1.8
-    targetCompatibility = 1.8
-    compileJava.options.encoding = 'UTF-8'
-    repositories {
-        jcenter()
-        mavenCentral()
-        maven { url = 'https://hub.spigotmc.org/nexus/content/repositories/snapshots' }
-        maven { url = 'https://oss.sonatype.org/content/repositories/snapshots' }
-        maven { url = 'https://repo.codemc.org/repository/nms/' }
-        maven { url = 'https://repo.codemc.org/repository/maven-public/' }
-        maven { url = 'http://repo.extendedclip.com/content/repositories/placeholderapi/' }
-        maven { url = 'https://jitpack.io' }
-    }
-    tasks.processResources {
-        filter ReplaceTokens, tokens: [
-                "version"    : project.property("version"),
-                "description": project.property("description")
-        ]
-    }
+group = 'net.leomixer17'
+version = '1.4.0'
+description = 'Create cool interactive books!'
+
+sourceCompatibility = 1.8
+targetCompatibility = 1.8
+
+tasks.compileJava.options.encoding = 'UTF-8'
+
+repositories {
+    jcenter()
+    mavenCentral()
+
+    maven { url = 'https://hub.spigotmc.org/nexus/content/repositories/snapshots/' }
+    maven { url = 'https://oss.sonatype.org/content/repositories/snapshots/' }
+    maven { url = 'http://repo.extendedclip.com/content/repositories/placeholderapi/' }
+    maven { url = 'https://repo.codemc.org/repository/maven-public/' }
+    maven { url = 'https://jitpack.io' }
 }
 
 dependencies {
-    compile project(path: ':core', configuration: 'shadow')
-    compile project(path: ':spigot-v1_8_R2', configuration: 'shadow')
-    compile project(path: ':spigot-v1_8_R3', configuration: 'shadow')
-    compile project(path: ':spigot-v1_9_R1', configuration: 'shadow')
-    compile project(path: ':spigot-v1_9_R2', configuration: 'shadow')
-    compile project(path: ':spigot-v1_10_R1', configuration: 'shadow')
-    compile project(path: ':spigot-v1_11_R1', configuration: 'shadow')
-    compile project(path: ':spigot-v1_12_R1', configuration: 'shadow')
-    compile project(path: ':spigot-v1_13_R1', configuration: 'shadow')
-    compile project(path: ':spigot-v1_13_R2', configuration: 'shadow')
-    compile project(path: ':spigot-v1_14_R1', configuration: 'shadow')
+    compileOnly 'org.spigotmc:spigot-api:1.14.4-R0.1-SNAPSHOT'
+    compileOnly 'net.md-5:bungeecord-chat:1.14-SNAPSHOT'
+    compileOnly 'me.clip:placeholderapi:2.10.3'
+    compileOnly group: 'io.netty', name: 'netty-all', version: '4.1.39.Final'
+    compile 'org.bstats:bstats-bukkit-lite:1.5'
+    compile 'com.github.Leomixer17.PluginLib:bukkit:-SNAPSHOT'
+}
+
+tasks.processResources {
+    filter ReplaceTokens, tokens: [
+            "version"    : project.property("version"),
+            "description": project.property("description")
+    ]
 }
 
 tasks.jar.enabled = false
 
 tasks.shadowJar {
     archiveFileName = 'InteractiveBooks.jar'
+    minimize()
+    relocate 'net.leomixer17.pluginlib', 'net.leomixer17.interactivebooks'
+    relocate 'org.bstats.bukkit', 'net.leomixer17.interactivebooks.bstats'
 }
 
 artifacts {
