--- core/build.gradle
@@ -1,24 +1,0 @@
-dependencies {
-    compileOnly 'org.spigotmc:spigot-api:1.14.4-R0.1-SNAPSHOT'
-    compileOnly 'net.md-5:bungeecord-chat:1.14-SNAPSHOT'
-    compileOnly 'me.clip:placeholderapi:2.10.3'
-    compile 'org.bstats:bstats-bukkit-lite:1.5'
-    compile 'com.github.Leomixer17.PluginLib:bukkit:-SNAPSHOT'
-}
-
-tasks.jar.enabled = false
-
-tasks.shadowJar {
-    archiveFileName = 'InteractiveBooks-Core.jar'
-    minimize()
-    relocate 'net.leomixer17.pluginlib', 'net.leomixer17.interactivebooks'
-    relocate 'org.bstats.bukkit', 'net.leomixer17.interactivebooks.bstats'
-}
-
-artifacts {
-    archives shadowJar
-}
-
-configurations.all {
-    transitive = false
-}
