+++ settings.gradle
@@ -1,14 +1,1 @@
 rootProject.name = 'InteractiveBooks'
-
-include 'core'
-
-include 'spigot-v1_8_R2'
-include 'spigot-v1_8_R3'
-include 'spigot-v1_9_R1'
-include 'spigot-v1_9_R2'
-include 'spigot-v1_10_R1'
-include 'spigot-v1_11_R1'
-include 'spigot-v1_12_R1'
-include 'spigot-v1_13_R1'
-include 'spigot-v1_13_R2'
-include 'spigot-v1_14_R1'
