--- spigot-v1_11_R1/build.gradle
@@ -1,4 +1,0 @@
-dependencies {
-    compileOnly 'org.spigotmc:spigot:1.11-R0.1-SNAPSHOT'
-    compileOnly project(path: ':core', configuration: 'shadow')
-}
