--- spigot-v1_13_R2/build.gradle
@@ -1,4 +1,0 @@
-dependencies {
-    compileOnly 'org.spigotmc:spigot:1.13.2-R0.1-SNAPSHOT'
-    compileOnly project(path: ':core', configuration: 'shadow')
-}
