--- spigot-v1_14_R1/build.gradle
@@ -1,4 +1,0 @@
-dependencies {
-    compileOnly 'org.spigotmc:spigot:1.14.3-R0.1-SNAPSHOT'
-    compileOnly project(path: ':core', configuration: 'shadow')
-}
