+++ src/main/java/net/leomixer17/interactivebooks/nms/BaseComponentSerializer.java
